// src/app/api/fn/customers/list-customers.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { CustomerList } from '../../models/customer-list';

export interface ListCustomers$Params {

/**
 * Case-insensitive substring match against customer name.
 */
  search?: string;
  page?: number;
  per_page?: number;
}

export function listCustomers(http: HttpClient, rootUrl: string, params?: ListCustomers$Params, context?: HttpContext): Observable<StrictHttpResponse<CustomerList>> {
  const rb = new RequestBuilder(rootUrl, listCustomers.PATH, 'get');
  if (params) {
    rb.query('search', params.search, {});
    rb.query('page', params.page, {});
    rb.query('per_page', params.per_page, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<CustomerList>;
    })
  );
}

listCustomers.PATH = '/customers';